Cleaned and split up naming and code based on testing
