+++ scripts/ci/roles/buildbot-master/files/src/golem_buildbot/services.py
@@ -6,11 +6,10 @@ from .settings import local_settings
 services = [
     reporters.GitHubStatusPush(
         builders=[
-            'unittest-fast_control',
-            'unittest_control',
-            'buildpackage_control'],
+            'control_test',
+            'control_build'],
         token=local_settings['github_api_token'],
-        context=util.Interpolate('buildbot/%(prop:buildtype)s'),
+        context=util.Interpolate('buildbot/%(prop:buildername)s'),
         startDescription='Build started.',
         endDescription='Build done.')
 ]
